refactor : 메소드 수정 및 수익률 계산 메소드 추가

+++ src/main/java/lotto/UserOutput.java
@@ -43,7 +43,7 @@ public class UserOutput {
         }
     }
 
-    public static void printWinningRate(float rate){
+    public static void printWinningRate(double rate){
         System.out.printf(WINNING_RATE_MESSAGE,rate);
     }
 }
